Replace joox with jxpath to improve performance
