+++ joylive-implement/joylive-parser/pom.xml
@@ -16,7 +16,7 @@
         <module>joylive-parser-fastjson2</module>
         <module>joylive-parser-jackson</module>
         <module>joylive-parser-properties</module>
-        <module>joylive-parser-joox</module>
+        <module>joylive-parser-jxpath</module>
     </modules>
 
 </project>
