+++ joylive-package/pom.xml
@@ -67,7 +67,7 @@
         </dependency>
         <dependency>
             <groupId>com.jd.live</groupId>
-            <artifactId>joylive-parser-joox</artifactId>
+            <artifactId>joylive-parser-jxpath</artifactId>
         </dependency>
         <dependency>
             <groupId>com.jd.live</groupId>
